requirements_typing.txt
mypy
types-pytz
django-stubs
